feat(catalog): Sku 엔티티와 리포지토리 추가
- 식별자 고갈 여지를 없애기 위해 id 계열 컬럼을 bigint로 상향
- JPA가 요구하는 무파라미터 생성자를 위해 모든 프로퍼티에 기본값 부여

+++ src/main/resources/db/migration/V1__create_sku_and_order.sql
@@ -1,14 +1,14 @@
 create table sku
 (
-    id    int primary key auto_increment,
+    id    bigint primary key auto_increment,
     name  varchar(255) not null,
     stock int          not null
 );
 
 create table orders
 (
-    id       int primary key auto_increment,
-    user_id  int not null,
-    sku_id   int not null,
-    quantity int not null
+    id       bigint primary key auto_increment,
+    user_id  bigint not null,
+    sku_id   bigint not null,
+    quantity int    not null
 );
